dark mode tv and uter stuff

+++ install.sh
@@ -5,28 +5,28 @@ sudo apt upgrade -y
 
 # Surface
 if [ "$(dmidecode -s system-product-name | grep -i Surface)" ]; then
-    # sudo apt install -y extrepo
-    # sudo extrepo enable surface-linux && sudo apt update
-    # sudo apt update -y
-    # sudo apt install -y linux-image-surface linux-headers-surface libwacom-surface iptsd
+    sudo apt install -y extrepo
+    sudo extrepo enable surface-linux && sudo apt update
+    sudo apt update -y
+    sudo apt install -y linux-image-surface linux-headers-surface libwacom-surface iptsd
 fi
 
 
-sudo snap install slack
-
 # Docker
-sudo apt-get update
-sudo apt-get install ca-certificates curl gnupg
+sudo apt-get update -y
+sudo apt-get install -y ca-certificates curl gnupg
 sudo install -m 0755 -d /etc/apt/keyrings
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
 sudo chmod a+r /etc/apt/keyrings/docker.gpg
+if [ ! -f /etc/apt/keyrings/docker.gpg ]; then
 echo \
-  "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
-  "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | \
-  sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+    "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
+    "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | \
+    sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+fi
 sudo apt update -y
-sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
-
+sudo apt install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+# File '/etc/apt/keyrings/docker.gpg' exists. Overwrite? (y/N)
 
 # Install Packages
 #paru -S --noconfirm --needed zsh arandr remmina-plugin-rdesktop docker sshpass remmina ansible qbittorrent gnu-netcat qemu-full networkmanager-l2tp networkmanager-strongswan remmina-plugin-ultravnc screen meld betterlockscreen_rapid-git dnsmasq rclone ntfs-3g flameshot acpid bc numlockx spotify-launcher unzip usbutils dmidecode autorandr pavucontrol variety termite feh git tree virt-manager dunst xclip xorg-xkill rofi acpilight nautilus scrot teamviewer network-manager-applet xautolock man powertop networkmanager nm-connection-editor network-manager-applet openvpn slack-desktop wget python google-chrome freecad gparted peak-linux-headers kicad i3exit polybar parsec-bin can-utils visual-studio-code-bin ttf-nerd-fonts-symbols libreoffice-fresh gnome-keyring subversion
@@ -35,6 +35,9 @@ sudo apt install -y zsh remmina ansible qbittorrent \
  virt-manager dunst xclip  rofi scrot powertop openvpn \
   wget python3 freecad gparted kicad polybar  
 
+sudo snap install slack
+
+# Vscode extensions
 code --install-extension alexcvzz.vscode-sqlite
 code --install-extension atlassian.atlascode 
 code --install-extension danielroedl.meld-diff eamodio.gitlens 
@@ -67,12 +70,20 @@ gsettings set org.gnome.desktop.interface color-scheme prefer-dark
 echo 'SUBSYSTEM=="backlight",RUN+="/bin/chmod 666 /sys/class/backlight/%k/brightness /sys/class/backlight/%k/bl_power"' | sudo tee /etc/udev/rules.d/backlight-permissions.rules
 sudo sh -c 'echo SUBSYSTEM=="drm", ACTION=="change", RUN+="/usr/bin/autorandr" > /etc/udev/rules.d/70-monitor.rules'
 
+# Download TeamViewer package
+wget https://download.teamviewer.com/download/linux/teamviewer_amd64.deb
+sudo dpkg -i teamviewer_amd64.deb
+sudo apt install -f -y
+rm teamviewer_amd64.deb
+# Darkmode TeamViewer
+if ! grep -q "ColorScheme = 2" $HOME/.config/teamviewer/client.conf; then
+    echo "[int32] ColorScheme = 2" >> $HOME/.config/teamviewer/client.conf
+fi
+
 # Enable services
 if ! systemctl is-active --quiet teamviewerd  ; then
     systemctl enable teamviewerd.service --now
 fi
-
-
 
 # sudo sh -c "echo blacklist nouveau > /etc/modprobe.d/blacklist-nvidia-nouveau.conf"
 # sudo sh -c "echo options nouveau modeset=0 >> /etc/modprobe.d/blacklist-nvidia-nouveau.conf"
@@ -115,7 +126,7 @@ if [ $USER = fw ]; then
 
     # paru -S --noconfirm --needed dwm st xorg-xinit xorg-server neovim rsync microsoft-edge-stable-bin qelectrotech libva-intel-driver dmenu prusa-slicer xidlehook
 
-elif [ $USER = user ]; then
+elif [ $USER = ingar ]; then
     git_url="https://github.com/ingar195/.dotfiles.git"
     
    
@@ -176,6 +187,8 @@ fi
 
 # Power settings
 sudo powertop --auto-tune
+sudo apt autoremove -y
+sudo apt autoclean -y
 
 echo ----------------------
 echo "Please reboot your PC"
